+++ react/components/EXPERIMENTAL_Table/SimpleTable/Header.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react'
 
 import Cell from './Cell'
-import { TABLE_HEADER_HEIGHT } from '../constants'
+import { TABLE_HEADER_HEIGHT, NAMESPACES } from '../constants'
 import useTableContext from '../hooks/useTableContext'
 
 const Header: FC = () => {
@@ -10,8 +10,9 @@ const Header: FC = () => {
   const renderHeader = (headerData: Column, headerIndex: number) => {
     const { headerRender, title, width } = headerData
     const content = headerRender ? headerRender({ headerData }) : title
+    const namespace = `header-${headerIndex}`
     return (
-      <Cell key={`col-${headerIndex}`} width={width} isHeader>
+      <Cell id={namespace} key={namespace} width={width} isHeader>
         {content}
       </Cell>
     )
@@ -20,6 +21,7 @@ const Header: FC = () => {
   return (
     <div
       key="header"
+      id={NAMESPACES.HEADER}
       className="dt-row w-100 h-100 ph4 truncate overflow-x-hidden c-muted-2 f6"
       style={{
         height: TABLE_HEADER_HEIGHT,