+++ react/components/EXPERIMENTAL_Table/Toolbar/Container.tsx
@@ -1,24 +1,20 @@
 import React, { FC } from 'react'
-import PropTypes, { InferProps } from 'prop-types'
+import { NAMESPACES, JUSTIFY_OPTIONS } from '../constants'
 
-const propTypes = {
-  justify: PropTypes.oneOf(['between', 'end', 'start', 'around', 'center']),
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node,
-  ]),
+type Props = {
+  justify?: FlexJustify
 }
 
-type Props = InferProps<typeof propTypes>
-
-const Container: FC<Props> = ({ justify = 'end', children }) => (
+const Container: FC<Props> = ({ justify, children }) => (
   <div
-    id="toolbar"
-    className={`vtex-tablev2_toolbar__container mb5 flex flex-row w-100 justify-${justify}`}>
+    id={NAMESPACES.TOOLBAR.CONTAINER}
+    className={`mb5 flex flex-row w-100 justify-${justify}`}>
     {children}
   </div>
 )
 
-Container.propTypes = propTypes
+Container.defaultProps = {
+  justify: JUSTIFY_OPTIONS.END,
+}
 
 export default Container