+++ react/components/EXPERIMENTAL_Table/Toolbar/Menu/Box.tsx
@@ -1,9 +1,10 @@
 import React, { FC } from 'react'
+import { BOX_ALIGNMENT } from '../../constants'
 
 export type BoxProps = {
   height?: string | number
   width?: string | number
-  alignMenu?: 'right' | 'left'
+  alignMenu?: Alignment
   noMargin?: boolean
   borderClasses?: string
 }
@@ -16,10 +17,11 @@ const Box: FC<BoxProps> = ({
   borderClasses,
   children,
 }) => {
+  const isAlignRight = alignMenu === BOX_ALIGNMENT.RIGHT
   return (
     <div
       className={`absolute z-999 shadow-4 ${
-        alignMenu === 'right' ? 'right-0' : 'left-0'
+        isAlignRight ? 'right-0' : 'left-0'
       } ${borderClasses || 'b--muted-4 br2 ba'} ${noMargin ? '' : 'mt2 mh2'}`}
       style={{
         width: width,