+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.73.14] - 2019-08-28
+
+### Added
+
+- `ButtonDownload` to `EXPERIMENTAL_TableV2`
+- `ButtonUpload` to `EXPERIMENTAL_TableV2`
+- `ButtonExtraActions` to `EXPERIMENTAL_TableV2`
+- `ButtonNewLine` to `EXPERIMENTAL_TableV2`
+
 ## [9.73.13] - 2019-08-27
 
 ### Fixed
@@ -24,7 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Add a class to Button component to fix its label alignment when using `href` and `block` props.
-
 
 ## [9.73.10] - 2019-08-23